ib._ledger: move trades transaction processing helpers into new module
