+++ piker/brokers/ib/__init__.py
@@ -35,6 +35,8 @@ from .feed import (
 )
 from .broker import (
     open_trade_dialog,
+)
+from .ledger import (
     norm_trade_records,
 )
